remove floating-link
